🚅 perfs: split psbtTransaction and ts types
