+++ src/psbt.js
@@ -13,6 +13,7 @@ const bscript = require('./script');
 const transaction_1 = require('./transaction');
 const bip371_1 = require('./psbt/bip371');
 const psbtutils_1 = require('./psbt/psbtutils');
+const transaction_2 = require('./psbt/transaction');
 /**
  * These are the default arguments for a Psbt instance.
  */
@@ -76,12 +77,18 @@ class Psbt {
     return this.fromBuffer(buffer, opts);
   }
   static fromBuffer(buffer, opts = {}) {
-    const psbtBase = bip174_1.Psbt.fromBuffer(buffer, transactionFromBuffer);
+    const psbtBase = bip174_1.Psbt.fromBuffer(
+      buffer,
+      transaction_2.transactionFromBuffer,
+    );
     const psbt = new Psbt(opts, psbtBase);
     checkTxForDupeIns(psbt.__CACHE.__TX, psbt.__CACHE);
     return psbt;
   }
-  constructor(opts = {}, data = new bip174_1.Psbt(new PsbtTransaction())) {
+  constructor(
+    opts = {},
+    data = new bip174_1.Psbt(new transaction_2.PsbtTransaction()),
+  ) {
     this.data = data;
     // set defaults
     this.opts = Object.assign({}, DEFAULT_OPTS, opts);
@@ -919,61 +926,6 @@ class Psbt {
   }
 }
 exports.Psbt = Psbt;
-/**
- * This function is needed to pass to the bip174 base class's fromBuffer.
- * It takes the "transaction buffer" portion of the psbt buffer and returns a
- * Transaction (From the bip174 library) interface.
- */
-const transactionFromBuffer = buffer => new PsbtTransaction(buffer);
-/**
- * This class implements the Transaction interface from bip174 library.
- * It contains a bitcoinjs-lib Transaction object.
- */
-class PsbtTransaction {
-  constructor(buffer = Buffer.from([2, 0, 0, 0, 0, 0, 0, 0, 0, 0])) {
-    this.tx = transaction_1.Transaction.fromBuffer(buffer);
-    checkTxEmpty(this.tx);
-    Object.defineProperty(this, 'tx', {
-      enumerable: false,
-      writable: true,
-    });
-  }
-  getInputOutputCounts() {
-    return {
-      inputCount: this.tx.ins.length,
-      outputCount: this.tx.outs.length,
-    };
-  }
-  addInput(input) {
-    if (
-      input.hash === undefined ||
-      input.index === undefined ||
-      (!Buffer.isBuffer(input.hash) && typeof input.hash !== 'string') ||
-      typeof input.index !== 'number'
-    ) {
-      throw new Error('Error adding input.');
-    }
-    const hash =
-      typeof input.hash === 'string'
-        ? (0, bufferutils_1.reverseBuffer)(Buffer.from(input.hash, 'hex'))
-        : input.hash;
-    this.tx.addInput(hash, input.index, input.sequence);
-  }
-  addOutput(output) {
-    if (
-      output.script === undefined ||
-      output.value === undefined ||
-      !Buffer.isBuffer(output.script) ||
-      typeof output.value !== 'number'
-    ) {
-      throw new Error('Error adding output.');
-    }
-    this.tx.addOutput(output.script, output.value);
-  }
-  toBuffer() {
-    return this.tx.toBuffer();
-  }
-}
 function canFinalize(input, script, scriptType) {
   switch (scriptType) {
     case 'pubkey':
@@ -1066,18 +1018,6 @@ function checkScriptForPubkey(pubkey, script, action) {
     throw new Error(
       `Can not ${action} for this input with the key ${pubkey.toString('hex')}`,
     );
-  }
-}
-function checkTxEmpty(tx) {
-  const isEmpty = tx.ins.every(
-    input =>
-      input.script &&
-      input.script.length === 0 &&
-      input.witness &&
-      input.witness.length === 0,
-  );
-  if (!isEmpty) {
-    throw new Error('Format Error: Transaction ScriptSigs are not empty');
   }
 }
 function checkTxForDupeIns(tx, cache) {
