+++ src/psbt/transaction.js
@@ -1,0 +1,74 @@
+'use strict';
+Object.defineProperty(exports, '__esModule', { value: true });
+exports.transactionFromBuffer = exports.PsbtTransaction = void 0;
+const bufferutils_1 = require('../bufferutils');
+const transaction_1 = require('../transaction');
+/**
+ * This class implements the Transaction interface from bip174 library.
+ * It contains a bitcoinjs-lib Transaction object.
+ */
+class PsbtTransaction {
+  constructor(buffer = Buffer.from([2, 0, 0, 0, 0, 0, 0, 0, 0, 0])) {
+    this.tx = transaction_1.Transaction.fromBuffer(buffer);
+    this.checkTxEmpty(this.tx);
+    Object.defineProperty(this, 'tx', {
+      enumerable: false,
+      writable: true,
+    });
+  }
+  getInputOutputCounts() {
+    return {
+      inputCount: this.tx.ins.length,
+      outputCount: this.tx.outs.length,
+    };
+  }
+  addInput(input) {
+    if (
+      input.hash === undefined ||
+      input.index === undefined ||
+      (!Buffer.isBuffer(input.hash) && typeof input.hash !== 'string') ||
+      typeof input.index !== 'number'
+    ) {
+      throw new Error('Error adding input.');
+    }
+    const hash =
+      typeof input.hash === 'string'
+        ? (0, bufferutils_1.reverseBuffer)(Buffer.from(input.hash, 'hex'))
+        : input.hash;
+    this.tx.addInput(hash, input.index, input.sequence);
+  }
+  addOutput(output) {
+    if (
+      output.script === undefined ||
+      output.value === undefined ||
+      !Buffer.isBuffer(output.script) ||
+      typeof output.value !== 'number'
+    ) {
+      throw new Error('Error adding output.');
+    }
+    this.tx.addOutput(output.script, output.value);
+  }
+  toBuffer() {
+    return this.tx.toBuffer();
+  }
+  checkTxEmpty(tx) {
+    const isEmpty = tx.ins.every(
+      input =>
+        input.script &&
+        input.script.length === 0 &&
+        input.witness &&
+        input.witness.length === 0,
+    );
+    if (!isEmpty) {
+      throw new Error('Format Error: Transaction ScriptSigs are not empty');
+    }
+  }
+}
+exports.PsbtTransaction = PsbtTransaction;
+/**
+ * This function is needed to pass to the bip174 base class's fromBuffer.
+ * It takes the "transaction buffer" portion of the psbt buffer and returns a
+ * Transaction (From the bip174 library) interface.
+ */
+const transactionFromBuffer = buffer => new PsbtTransaction(buffer);
+exports.transactionFromBuffer = transactionFromBuffer;
